requirements.txt
langchain==0.3.15
langchain-community==0.3.15
faiss-cpu==1.10.0
pypdf==5.1.0
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers
langchain-huggingface
groq